requirements.txt
opencv-python~=4.5.1.48
numpy~=1.20.2
python-interface~=1.6.1